+++ packages/react/package.json
@@ -10,7 +10,7 @@
     "format": "prettier --write './**/*.{ts,tsx,html,json}'",
     "format-check": "prettier --check './**/*.{ts,tsx,html,json}'",
     "lint": "eslint .",
-    "build:css": "mkdir -p dist && npx tailwindcss -o ./dist/styles.css --minify"
+    "build:css": "mkdir -p dist && npx tailwindcss -i ./src/index.css -o ./dist/index.css --minify"
   },
   "dependencies": {
     "@czi-sds/components": "^22.1.2",
