+++ packages/react/src/components/Main.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
+import '../index.css';
 
 import ThemedApp from "./ThemedApp.tsx";
